travis and linter added
