--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-react",
-    "@babel/preset-env"
-    ],
-    /*
-    Babel uses these "presets" to know how to transpile your code:
-      'react': teaches Babel to recognize JSX
-      'env': teaches Babel to transpile Javascript. You can reduce the size of your bundle by limiting the number of features/plug-ins you transpile. Learn more at: https://github.com/babel/babel-preset-env
-    */
-    "plugins": [
-      "@babel/plugin-transform-runtime"
-    ]
-}
