+++ tests/test.spec.js
@@ -1,0 +1,7 @@
+
+//test to test that the test suite is working
+describe('test suite', () => {
+  it('returns true', () => {
+    expect(true).toEqual(true);
+  });
+});
